MAINT: Moved common parts of NewtonSolver and LinearSolver into a separate file
